pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
flask = "==1.0.2"
python = "^3.7"
Flask-SQLAlchemy = "^2.4"
Flask-JWT = "^0.3.2"
Flask-WTF = "^0.14.3"
Flask-Login = "^0.5.0"
